test(ttsx): align the runtime suite with the per-file emit model
The byte-store / per-package dependency-build subsystem was removed in favour of
ts-node-style per-run in-memory emit. These tests asserted that gone machinery;
they now verify the surviving behaviour and drop the removed-internal checks:

- type-check gate, keep-outdir, relative-cache-dir: drop the per-run
  `<cacheDir>/project` scratch assertions; keep the user-visible behaviour.
- emits-a-raw-dependency-fresh-on-every-run (was reuses-and-rebuilds-cache):
  asserts edits are reflected and unchanged sources are stable, with no cache.
- loose-emits-a-dependency-file-its-package-tsconfig-excludes (was
  reports-a-clear-error...emits-nothing): such a file now loose-compiles and
  runs instead of erroring.
- runs-concurrent-first-emits (was handles-concurrent-first-builds-cache): keeps
  the concurrency behaviour, drops cache stamp/scratch assertions.
- uses-the-consumer-local/explicit-tsgo-for-the-type-check (was
  ...-emitted-by/...-for-dependency-builds): the supplied tsgo runs the
  `--noEmit` check; emission is the driver host's job.

Removed (contract structurally void in the new model):
- reaps-a-dead-pid-scratch-directory: no promotion scratch exists.
- rebuilds-cache-when-extends-base-changes: no cache hash to miss.
- rebuilds-cache-when-{explicit,auto}-plugin-file-changes: a dependency cannot
  declare its own divergent plugins; deps emit through the entry's single host.
