+++ tests/test-ttsc/src/features/ttsx-runtime/test_runner_corpus_ttsx_keeps_configured_outdir_untouched.ts
@@ -49,7 +49,4 @@ export const test_runner_corpus_ttsx_keeps_configured_outdir_untouched = () => {
   );
   assert.equal(fs.existsSync(path.join(root, "dist", "main.js")), false);
   assert.equal(fs.existsSync(path.join(root, "dist", "package.json")), false);
-  const projectCache = path.join(cacheDir, "project");
-  assert.equal(fs.existsSync(projectCache), true);
-  assert.deepEqual(fs.readdirSync(projectCache), []);
 };
