+++ tests/test-ttsc/src/features/ttsx-runtime/test_ttsx_runs_concurrent_first_emits_of_a_raw_ts_dependency.ts
@@ -1,26 +1,22 @@
 import { TestProject } from "@ttsc/testing";
 import assert from "node:assert/strict";
 import child_process from "node:child_process";
-import fs from "node:fs";
-import path from "node:path";
 
 /**
- * Verifies `ttsx` handles concurrent first builds of one raw-`.ts` dependency.
+ * Verifies `ttsx` runs several concurrent first emits of one raw `.ts`
+ * dependency without interfering.
  *
- * The dependency cache is shared by every `ttsx` process running the same
- * project. A cold cache therefore has a real promotion race: several processes
- * can compile into private staging directories before any complete cache
- * exists. The winner must publish a complete cache, the losers must keep using
- * a valid cache without disturbing it, and no staging/retired directories may
- * be left behind.
+ * Every `ttsx` process compiles the dependency through its own persistent host,
+ * so a cold start has no shared promotion to race over: each process emits the
+ * dependency independently in memory. Launching many at once against the same
+ * cold project must leave every one with correct output and a clean tree.
  *
  * 1. Create an ESM project whose raw dependency imports many TypeScript files and
- *    uses a non-erasable enum, forcing a real dependency build.
+ *    uses a non-erasable enum, forcing a real dependency compile.
  * 2. Launch four `ttsx` processes against the same cold project at once.
- * 3. Assert every process exits zero, the dependency cache is complete, and no
- *    promotion scratch directories remain beside it.
+ * 3. Assert every process exits zero with the expected output.
  */
-export const test_ttsx_handles_concurrent_first_builds_of_a_raw_ts_dependency_cache =
+export const test_ttsx_runs_concurrent_first_emits_of_a_raw_ts_dependency =
   async () => {
     const files: Record<string, string> = {
       "package.json": JSON.stringify({ type: "module", private: true }),
@@ -86,29 +82,6 @@ export const test_ttsx_handles_concurrent_first_builds_of_a_raw_ts_dependency_ca
       assert.equal(result.status, 0, result.stderr);
       assert.equal(result.stdout.trim(), `${result.label}:race-ok:2016`);
     }
-
-    const depRoot = path.join(root, "node_modules", "race-dep");
-    const cacheDir = path.join(
-      depRoot,
-      "node_modules",
-      ".cache",
-      "ttsc",
-      "ttsx-deps",
-    );
-    assert.ok(
-      fs.existsSync(path.join(cacheDir, ".ttsx-stamp.json")),
-      "the promoted dependency cache is stamped as complete",
-    );
-    assert.notEqual(
-      findEmittedEntry(depRoot),
-      null,
-      "the dependency entry was emitted into the completed cache",
-    );
-    assert.deepEqual(
-      cacheScratchEntries(cacheDir),
-      [],
-      "promotion scratch directories are cleaned after the race",
-    );
   };
 
 function launchTtsx(
@@ -146,44 +119,4 @@ function launchTtsx(
     child.on("error", reject);
     child.on("close", (status) => resolve({ label, status, stdout, stderr }));
   });
-}
-
-function findEmittedEntry(packageRoot: string): string | null {
-  const stack = [
-    path.join(packageRoot, "node_modules", ".cache", "ttsc", "ttsx-deps"),
-  ];
-  while (stack.length !== 0) {
-    const current = stack.pop()!;
-    let entries: fs.Dirent[];
-    try {
-      entries = fs.readdirSync(current, { withFileTypes: true });
-    } catch {
-      continue;
-    }
-    for (const entry of entries) {
-      const next = path.join(current, entry.name);
-      if (entry.isDirectory()) {
-        stack.push(next);
-      } else if (entry.isFile() && entry.name === "index.js") {
-        return next;
-      }
-    }
-  }
-  return null;
-}
-
-function cacheScratchEntries(cacheDir: string): string[] {
-  const parent = path.dirname(cacheDir);
-  try {
-    return fs
-      .readdirSync(parent)
-      .filter(
-        (entry) =>
-          entry.startsWith("ttsx-deps.") &&
-          (entry.endsWith(".staging") || entry.endsWith(".retired")),
-      )
-      .sort();
-  } catch {
-    return [];
-  }
 }
